What does this do?
I was looking at two cars and wanted to see in terms of the monthly car note, the gas, and the commute which of the two would end up costing more in the long-run just off of these two variables alone. 
The script is based on inputs from the user and it will do the math for you on those inputs to return which car would be the cheapest option between the two. 

--
Example:
Lexus with 16/21 MPG at $500 per month. 
vs
BMW with 22/25 MPG at $750 per month. 
Both on the same estimated commute which would cost more to own?